+++ vault/CHANGELOG.md
@@ -1,10 +1,15 @@
+2.0
+
+* Updated to use local consol agent and a consul cluster for data store
+
+---
+
 1.0.1
 
-* Rebuild for FreeBSD 13 & new packages
+* First Vault pot image
 
 ---
 
 1.0
 
-* First Vault pot image
-
+* initiate file